refactor: clean routefiles exporting factories to factories folder
